fix(exec): always new pid ns in third layer box
As the 'exec' function is implemented by forking a new child from the
init process. The init process have to stay in the same user namespace
as application process.

So we now result in a situation where we have to always create new pid
namespace to keep the third layer box as the init(pid=1) process for
application.

Change-Id: If83dd5bc10c292f43b9ad3e0fb59b31d222fccac